requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
azure-core==1.40.0
azure-cosmos==4.15.0
boto3==1.43.2
botocore==1.43.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
fastapi==0.136.1
google-api-core==2.30.3
google-auth==2.50.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
idna==3.13
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pika==1.3.2
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
requests==2.33.1
s3transfer==0.17.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
